+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project roughly adheres to [Semantic Versioning](https://semver.org/spe
 ### Added
 - **New type:** There is now a `JSONColumnType` using Arrow's support for structured data. This is ideal for multi-turn conversations etc.
 - **New op:** Text columns have a `.parse_json()` method.
+- **New op:** Text columns can be turned into boolean columns using `.is_in()` and `.not_in()`.
 - **Docs:** Improved docstrings for `.llm_layer_activations`
 - **Breaking:** `LLMLayerActivations` now supports a `apply_chat_template` parameter defaulting to True.